Refactor [#83] 조금 더 세분화된 클래스 형태로 리팩토링 진행
